backend/requirements_build.txt
# 运行依赖
Flask==2.3.3
Flask-CORS==4.0.0
pyserial==3.5
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
waitress==3.0.0

# 打包依赖
pyinstaller==6.3.0
pywebview==4.4.1

# PyVISA（如需真实设备）
# pyvisa==1.14.1
# pyvisa-py==0.7.1